requirements.txt
azure-ai-projects>=1.0.0b4
azure-identity>=1.15.0
azure-ai-agents>=1.0.0b2
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
